requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
pyyaml>=6.0
schedule>=1.2.0
jinja2>=3.1.0
openai>=1.0.0
